Refactor snappy functions into separate file
