+++ cubi_tk/snappy/kickoff.py
@@ -12,6 +12,7 @@ from toposort import toposort
 from cubi_tk.exceptions import ParseOutputException
 
 from . import common
+from .snappy_workflows import get_snappy_step_directories, get_workflow_step_dependencies
 
 
 def run(
@@ -26,9 +27,9 @@ def run(
     logger.info("Looking for pipeline directories (needs to contain snappy config.yaml)...")
     logger.debug("Looking in %s", path)
     step_set = {}
-    folder_steps = common.get_snappy_step_directories(path)
+    folder_steps = get_snappy_step_directories(path)
     for step_name, step_path in folder_steps.items():
-        dependencies = common.get_workflow_step_dependencies(step_path)
+        dependencies = get_workflow_step_dependencies(step_path)
         step_set[step_name] = dependencies
 
     steps: typing.List[str] = []
